Fix multitrack view bug with regions at the end of the track
Problem: When adding regions at the end of a track in the multitrack view,
if the length of the content within the region is longer than the width of
the region itself, the zoom level gets adjusted to fit the entire content
within the container.

Solution: Improved the text alignment within regions. When the duration of
a region is shorter than 0.5 seconds, the text content is now aligned to
the right side of the region to prevent the zoom level from being adjusted.